requirements.txt
streamlit>=1.32.0
google-api-python-client>=2.111.0
prophet>=1.1.5
scikit-learn>=1.4.0
pandas>=2.2.0
plotly>=5.18.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
google-generativeai>=0.8.0
